% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classVenn.R
\name{all_identical}
\alias{all_identical}
\title{All members of a list have the same elements}
\usage{
all_identical(list)
}
\arguments{
\item{list}{a list}
}
\value{
TRUE or FALSE
}
\description{
All members of a list have the same elements
}
