% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_shape.R
\name{build_shape}
\alias{build_shape}
\title{Helper function to add shape}
\usage{
build_shape(
  edge,
  label,
  nsets = length(edge),
  shape_id,
  type = c("ellipse", "triangle", "polygon", "circle")
)
}
\arguments{
\item{edge}{a list of xy matrix}

\item{label}{a list of xy matrix}

\item{nsets}{2:7}

\item{shape_id}{a unique id}

\item{type}{c("ellipse","triangle","polygon","circle")}
}
\value{
a tibble with columns: nsets, type, shape_id, component, id, xy.
}
\description{
Helper function to add shape
}
