% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{circle}
\alias{circle}
\title{generating a circle}
\usage{
circle(x, y, r, n = 1000)
}
\arguments{
\item{x, y}{center of circle}

\item{r}{radius of circle}

\item{n}{points (resolution)}
}
\value{
a data.frame representing circle position
}
\description{
generating a circle
}
