% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOperations.R
\name{discern,Polygon-method}
\alias{discern,Polygon-method}
\title{calculate the difference region of `Polygon` object}
\usage{
\S4method{discern}{Polygon}(venn, slice1, slice2 = "all")
}
\arguments{
\item{venn}{Venn/Polygon object}

\item{slice1}{first slice of Venn object}

\item{slice2}{second slice of Venn object, default is all except the first slice}
}
\description{
calculate the difference region of `Polygon` object
}
