% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{draw_4d_venn}
\alias{draw_4d_venn}
\title{draw 4d venn diagram}
\usage{
draw_4d_venn(x, n.sides, category.names, label, ...)
}
\arguments{
\item{x}{}

\item{n.sides}{}
}
\description{
draw 4d venn diagram
}
