% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2d_venn.R, R/draw_3d_venn.R,
%   R/draw_4d_venn.R
\name{draw_2d_venn}
\alias{draw_2d_venn}
\alias{draw_3d_venn}
\alias{draw_venn}
\alias{draw_4d_venn}
\title{draw 2d, 3d, and 4d Venn diagram}
\usage{
draw_2d_venn(x, n.sides, category.names, label, ...)

draw_3d_venn(x, n.sides, category.names, label, ...)

draw_4d_venn(x, n.sides, category.names, label, ...)
}
\arguments{
\item{x}{a list of items}

\item{n.sides}{resolution}

\item{category.names}{default is names(x)}

\item{label}{c("both","percent","count")}

\item{...}{passing to geom_polygon, enabling modification of polygon styles}
}
\description{
draw 2d, 3d, and 4d Venn diagram
}
