% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{fancy_2d_circle}
\alias{fancy_2d_circle}
\title{two dimension circle}
\usage{
fancy_2d_circle(parameters = NULL, n = 100)
}
\arguments{
\item{parameters}{will pass to shape generators}

\item{n}{count of points to shape this polygon}
}
\description{
two dimension circle
}
