% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{fancy_3d_circle}
\alias{fancy_3d_circle}
\title{fancy 3d circle}
\usage{
fancy_3d_circle(parameters = NULL, n = 100)
}
\arguments{
\item{parameters}{will pass to shape generators}

\item{n}{count of points to shape this polygon}
}
\description{
fancy 3d circle
}
