% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{fancy_4d_ellipse}
\alias{fancy_4d_ellipse}
\title{fancy 4d ellipse from `VennDiagram`}
\usage{
fancy_4d_ellipse(parameters = NULL, n = 100)
}
\arguments{
\item{parameters}{will pass to shape generators}

\item{n}{count of points to shape this polygon}
}
\value{
a list of coordinates matrix
}
\description{
fancy 4d ellipse from `VennDiagram`
}
