% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{fancy_6d_triangle}
\alias{fancy_6d_triangle}
\title{Six dimension triangle}
\usage{
fancy_6d_triangle(parameters = NULL)
}
\arguments{
\item{parameters}{will pass to shape generators}
}
\description{
Six dimension triangle
}
