% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_5d_venn.R
\name{five_dimension_ellipse_regions}
\alias{five_dimension_ellipse_regions}
\title{coordinations of polygon regions/centers for 4d venn diagram}
\usage{
five_dimension_ellipse_regions(n.sides)
}
\description{
coordinations of polygon regions/centers for 4d venn diagram
}
