% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_5d_venn.R
\name{five_dimension_region_values}
\alias{five_dimension_region_values}
\title{calculating count values}
\usage{
five_dimension_region_values(x)
}
\arguments{
\item{x}{a list of vector items.}
}
\description{
calculating count values
}
