% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{four_dimension_ellipse_regions}
\alias{four_dimension_ellipse_regions}
\title{coordinations of polygon regions/centers for 4d venn diagram}
\usage{
four_dimension_ellipse_regions(n.sides)
}
\arguments{
\item{n.sides}{}
}
\description{
coordinations of polygon regions/centers for 4d venn diagram
}
