% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{four_dimension_region_values}
\alias{four_dimension_region_values}
\title{calculating count values}
\usage{
four_dimension_region_values(x)
}
\arguments{
\item{x}{}
}
\description{
calculating count values
}
