% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{get_region_items}
\alias{get_region_items}
\title{get a list of items in Venn regions}
\usage{
get_region_items(x, category.names = names(x))
}
\arguments{
\item{x}{list of items}

\item{category.names}{default is names(x)}
}
\value{
a list of region items
}
\description{
get a list of items in Venn regions
}
