% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{get_section_df}
\alias{get_section_df}
\title{Function to generate a dataframe with the names of the items that make up the different sections}
\usage{
get_section_df(l, sep = ",")
}
\arguments{
\item{l}{a list (same you use to generate a venn diagram)}

\item{sep}{the seperator}
}
\value{
df
}
\description{
Function to generate a dataframe with the names of the items that make up the different sections
}
