% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape.R
\name{get_shape_by_id}
\alias{get_shape_by_id}
\title{Specifying a shape}
\usage{
get_shape_by_id(id)
}
\arguments{
\item{id}{shape id}
}
\value{
a shape
}
\description{
Specifying a shape
}
\examples{
get_shape_by_id("401f")
}
