% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\docType{package}
\name{ggVennDiagram-package}
\alias{ggVennDiagram-package}
\title{`ggVennDiagram`: an easy to use Venn diagram generator}
\description{
Venn diagram is frequently used in scientific studies of many fields.
This package incorporates state-of-art Venn plot tools and provides a set of
easy-to-use functions to plot Venn. By dealing with a user-provided list,
which contains the sets of Venn, `ggVennDiagram` returns a structured data that
can be used to plot Venn. The data contains three slots: 1) the edge of Venn sets;
2) the separated regions of Venn sets; 3) the labels of Venn sets.
By help from the package `venn`, it is possible to draw Venn diagram up to 7 sets.
}
