% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{interactive_venn}
\alias{interactive_venn}
\title{Plot interactive venn diagram}
\usage{
interactive_venn(
  l,
  linewidth = 0.1,
  linecolor = "black",
  labelsize = 8,
  countsize = 4,
  return_plotly = T
)
}
\arguments{
\item{l}{list with the different elements.}

\item{linewidth}{linewidth for the different sections of the diagram}

\item{linecolor}{color for the lines of the diagram}

\item{labelsize}{size of the labels (i.e. A or Ctrl)}

\item{countsize}{size of the labels inside the sections of the venn diagram}

\item{return_plotly}{Should the plotly diagram be returned or the ggplot?}
}
\value{

}
\description{
Plot interactive venn diagram
}
