% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{label_position}
\alias{label_position}
\title{helper function to set label position}
\usage{
label_position(position)
}
\arguments{
\item{position}{a data.frame containing label coordinates}
}
\value{
a list of matrix
}
\description{
helper function to set label position
}
