% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_implement.R
\name{multi}
\alias{multi}
\alias{multi_union}
\alias{multi_intersect}
\alias{multi_setdiff}
\title{Performs set union/intersection/diff on more than two vectors.}
\usage{
multi_union(..., l = NULL)

multi_intersect(..., l = NULL)

multi_setdiff(..., l = NULL)
}
\arguments{
\item{...}{at least three items are needed if use this parameter}

\item{l}{a list of vectors}
}
\description{
Performs set union/intersection/diff on more than two vectors.
}
