% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_implement.R
\name{multi_st_fun}
\alias{multi_st_fun}
\alias{st_multi_intersection}
\alias{st_multi_difference}
\alias{st_multi_union}
\title{Perform geometric set intersection, difference, and union with more than two simple feature geometry collections}
\usage{
st_multi_intersection(..., l = NULL)

st_multi_difference(..., l = NULL)

st_multi_union(..., l = NULL)
}
\arguments{
\item{...}{at least three items are needed if use this parameter}

\item{l}{a list of polygons}
}
\value{
intersection/union/diff of items
}
\description{
Perform geometric set intersection, difference, and union with more than two simple feature geometry collections
}
