% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetOperations.R
\name{overlap,Polygon-method}
\alias{overlap,Polygon-method}
\title{calculate the overlap region of `Polygon` object}
\usage{
\S4method{overlap}{Polygon}(venn, slice = "all")
}
\arguments{
\item{venn}{Venn object}

\item{slice}{a numeric vector indicating the index of slice, default is "all"}
}
\description{
calculate the overlap region of `Polygon` object
}
