% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{plot_shape_edge}
\alias{plot_shape_edge}
\title{Plot the set edge of a VennPlotData}
\usage{
plot_shape_edge(x)
}
\arguments{
\item{x}{a VennPlotData object}
}
\value{
a ggplot object
}
\description{
This is for viewing the shape id and appearance of the shape.
}
\examples{
  shape = get_shape_by_id("301")
  plot_shape_edge(shape)
}
