% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\name{plot_shapes}
\alias{plot_shapes}
\title{plot all shapes provided by internal dataset}
\usage{
plot_shapes()
}
\description{
These shapes are mainly collected from the package \code{venn}, and \code{VennDiagram}.
For Venn plot with more than 4 sets, it is usually impossible to plot with
simple circle or ellipse. So we need to use a predefined coordinates in plot.
}
\details{
\itemize{
\item Shape 101, 201, 301, 401, 402, 501, 502, 601 and 701 are from \code{venn}
\item Shape 401f is from \code{VennDiagram}
}

see \code{data-raw/shapes.R} to find how we incorporate these data.
}
\examples{
plot_shapes()
}
