% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{plot_venn}
\alias{plot_venn}
\title{plot codes}
\usage{
plot_venn(
  x,
  show_intersect = FALSE,
  label = "count",
  label_geom = "label",
  label_alpha = 0.3,
  percent_digit = 0,
  txtWidth = 40
)
}
\arguments{
\item{x}{list of items}

\item{show_intersect}{if TRUE the text can be visualized by `plotly`}

\item{label}{select one from c("count","percent","both","none")}

\item{label_geom}{choose from c("label", "text")}

\item{label_alpha}{set 0 to remove label background}

\item{percent_digit}{number of digits when formatting percent label (0)}

\item{txtWidth}{width of text used in showing intersect members (40)}
}
\value{
ggplot object, or plotly object if show_intersect is TRUE
}
\description{
plot codes
}
