% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{plotData_add_venn}
\alias{plotData_add_venn}
\title{join the shape data with set data}
\usage{
plotData_add_venn(plotData, venn)
}
\arguments{
\item{plotData}{a VennPlot object that stores plot shapes}

\item{venn}{a Venn object that stores set values}
}
\description{
join the shape data with set data
}
