% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\docType{methods}
\name{Polygon}
\alias{Polygon}
\alias{Polygon,ANY-method}
\title{Polygon constructor}
\usage{
Polygon(sets)

\S4method{Polygon}{ANY}(sets)
}
\arguments{
\item{sets}{a list containing multiple simple features}
}
\description{
Polygon constructor
}
