% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\name{Polygon}
\alias{Polygon}
\title{An S4 class to represent multiple polygons.}
\usage{
Polygon(sets)
}
\description{
An S4 class to represent multiple polygons.
}
\section{Slots}{

\describe{
\item{\code{sets}}{A `list` object contains sets}

\item{\code{names}}{The names of the `sets` if has names. If the `list`
doesn't have names, the sets will be named as "Set_1", "Set_2"
and so on.}
}}

