% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_shape.R
\docType{methods}
\name{print}
\alias{print}
\alias{print.VennPlotData}
\title{S3 method for \code{VennPlotData}}
\usage{
\method{print}{VennPlotData}(x, ...)
}
\arguments{
\item{x}{a VennPlotData object}

\item{...}{useless}
}
\description{
S3 method for \code{VennPlotData}
}
