% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{process_data}
\alias{process_data}
\title{get plot data}
\usage{
\S4method{process_data}{Venn}(venn, ...)
}
\arguments{
\item{venn}{a Venn object}

\item{...}{filter shapes. i.e. shape_id == "601", type == "polygon"}
}
\description{
get plot data
}
