% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{process_upset_data}
\alias{process_upset_data}
\title{process upset data}
\usage{
process_upset_data(
  venn,
  nintersects = 30,
  order.intersect.by = "size",
  order.set.by = "name",
  name_separator = "/"
)
}
\arguments{
\item{venn}{a class Venn object}

\item{nintersects}{number of intersects. If NULL, all intersections will show.}

\item{order.intersect.by}{'size', 'name', or "none"}

\item{order.set.by}{'size', 'name', or "none"}

\item{name_separator}{will be used to assign subset names}
}
\value{
a upsetPlotData object
}
\description{
process upset data
}
