% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2d_venn.R, R/draw_3d_venn.R,
%   R/draw_4d_venn.R
\name{two_dimension_region_items}
\alias{two_dimension_region_items}
\alias{three_dimension_region_items}
\alias{region_item}
\alias{four_dimension_region_items}
\title{get the items in each region}
\usage{
two_dimension_region_items(x)

three_dimension_region_items(x)

four_dimension_region_items(x)
}
\arguments{
\item{x}{a list of vector items}
}
\value{
a list
}
\description{
get the items in each region
}
