% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2d_venn.R, R/draw_3d_venn.R,
%   R/draw_4d_venn.R
\name{two_dimension_region_values}
\alias{two_dimension_region_values}
\alias{three_dimension_region_values}
\alias{region_value}
\alias{four_dimension_region_values}
\title{calculating intersection values of venn}
\usage{
two_dimension_region_values(x)

three_dimension_region_values(x)

four_dimension_region_values(x)
}
\arguments{
\item{x}{a list of vector items.}
}
\value{
data.frame
}
\description{
calculating intersection values of venn
}
