% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upset_plot.R
\name{separate_longer_delim}
\alias{separate_longer_delim}
\title{Implement of \code{tidyr::separate_longer_delim}}
\usage{
separate_longer_delim(df, col, delim)
}
\arguments{
\item{df}{a data.frame}

\item{col}{column}

\item{delim}{delimeter}
}
\value{
a data.frame
}
\description{
Implement of \code{tidyr::separate_longer_delim}
}
