% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_shapes.R
\docType{data}
\name{shapes}
\alias{shapes}
\title{shapes: shape data used to setup Venn plot}
\format{
a list with several slots
see "?VennPlotData".
}
\source{
\itemize{
\item The \code{venn} datasets authored by Adrian Dusa (\url{https://CRAN.R-project.org/package=venn}).
\item Parameters used to generate fancy four set ellipses are adopted from \code{VennDiagram}(\url{https://CRAN.R-project.org/package=VennDiagram}).
\item \href{https://upload.wikimedia.org/wikipedia/commons/5/56/6-set_Venn_diagram_SMIL.svg}{Wiki}
}
}
\description{
a collection of geometric shapes, which defined the edge and label of sets in a Venn plot.
use \code{plot_shapes()} to see some of them.
}
