% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classVenn.R
\name{show,Venn-method}
\alias{show,Venn-method}
\title{method for S4 Venn object}
\usage{
\S4method{show}{Venn}(object)
}
\arguments{
\item{object}{a Venn class object}
}
\description{
Print user-friendly information of a Venn object
}
