% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setOperations.R
\name{slice_idx}
\alias{slice_idx}
\title{check and format slice name}
\usage{
slice_idx(venn, slice)
}
\arguments{
\item{venn}{a Venn object}

\item{slice}{a numeric or character vector}
}
\value{
the index of Venn (numeric vector) or "all"
}
\description{
check and format slice name
}
