% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{triangle}
\alias{triangle}
\title{defined a triangle by three points}
\usage{
triangle(xy = c(0, 0, 1, 0, 0, 1))
}
\arguments{
\item{xy}{coordinates of the three points defining a triangle}
}
\value{
a matrix with xy coordinates
}
\description{
defined a triangle by three points
}
\examples{
# triangle coordinates
library(ggVennDiagram)
library(sf)
triangle()

# plot a new triangle
triangle(c(-1,0,1,0,0,2)) \%>\% st_linestring() \%>\% plot()
}
