% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classVenn.R
\name{Venn-method}
\alias{Venn-method}
\alias{Venn,ANY-method}
\title{Venn class object constructor}
\usage{
Venn(sets, names)

\S4method{Venn}{ANY}(sets, names = NULL)
}
\arguments{
\item{sets}{(Required) A list containing vectors in the same class. If a
vector contains duplicates they will be discarded. If the list doesn't have
names the sets will be named as "Set_1", "Set_2", "Set_3" and so on.}

\item{names}{names of sets}
}
\value{
A \code{Venn} object.
}
\description{
\code{Venn()} builds a \code{Venn} object from a list.
}
\examples{
 venn = Venn(list(letters[1:10], letters[3:12], letters[6:15]))
 print(venn)
}
