% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{venn_plot_data}
\alias{venn_plot_data}
\alias{venn_region}
\alias{venn_setedge}
\alias{venn_setlabel}
\title{Get VennPlotData slot}
\usage{
venn_region(obj)

venn_setedge(obj)

venn_setlabel(obj)
}
\arguments{
\item{obj}{a list that stores all the data from the S4 class `VennPlotData` object}
}
\value{
a tibble, `sf` object
}
\description{
Get VennPlotData slot
}
\examples{
venn_region(obj)   # return region data
venn_setlabel(obj) # return setLabel data
venn_setedge(obj)  # return setEdge data
}
