% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{vennplot-shapes}
\alias{vennplot-shapes}
\title{shapes: shape data used to setup Venn plot}
\format{
a list with several slots
 see "?vennPlotData".
}
\source{
- `venn:::sets`
- `library(VennDiagram)`
- [Wiki](https://upload.wikimedia.org/wikipedia/commons/5/56/6-set_Venn_diagram_SMIL.svg)
}
\description{
a collection of geometric shapes, which defined the edge and label of sets in a Venn plot.
use `plot_shapes()` to see some of them.
}
