% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\docType{class}
\name{VennPlotData-class}
\alias{VennPlotData-class}
\title{An S4 class to represent Venn plot components.}
\description{
An S4 class to represent Venn plot components.
}
\section{Slots}{

\describe{
\item{\code{setEdge}}{a list of coordinates matrix defining Venn set edges}

\item{\code{SetLabel}}{a list of coordinates matrix defining Venn set labels}

\item{\code{region}}{the feature region will be calculated automatically with `setEdge`}
}}

