% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_shape.R
\name{VennPlotData}
\alias{VennPlotData}
\title{An S3 class constructor of representing Venn plot components.}
\usage{
VennPlotData(x)
}
\arguments{
\item{x}{data source of a VennPlotData object}
}
\description{
An S3 class constructor of representing Venn plot components.
}
\section{Slots}{

\describe{
\item{\code{shapeId}}{shape id}

\item{\code{type}}{type of shape}

\item{\code{nsets}}{number of sets}

\item{\code{setEdge}}{a data.frame, the coordinates of set edges}

\item{\code{setLabel}}{a data.frame, the coordinates of set labels}

\item{\code{regionEdge}}{a data.frame, the coordinates of different regions}

\item{\code{regionLabel}}{a data.frame, the centroid of the regions, where region labels anchored}
}}

