% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{circle}
\alias{circle}
\title{generating a circle}
\usage{
circle(x = 0, y = 0, r = 1, n = 100)
}
\arguments{
\item{x, y}{center of circle}

\item{r}{radius of circle}

\item{n}{number of points for polygon object (resolution)}
}
\value{
a matrix representing circle coordinates
}
\description{
generating a circle
}
\examples{
# plot the default circle
library(ggVennDiagram)
library(sf)
circle() \%>\% st_linestring() \%>\% plot()
}
