% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{ellipse}
\alias{ellipse}
\title{generating a closed ellipse}
\usage{
ellipse(x = 0, y = 0, a = 2, b = 1, rotation = 0, n = 100)
}
\arguments{
\item{x, y}{the coordinates of ellipse center}

\item{a}{radius of short arm}

\item{b}{radius of long arm}

\item{rotation}{rotation in degree}

\item{n}{number of points}
}
\value{
a matrix representing ellipse coordinates
}
\description{
This function is derived from `VennDiagram::ell2poly`, we modified it and then
it can generating a closed ellipse, which is a requirement for further transformation
to a POLYGON sf object.
}
\examples{
# plot the default ellipse
library(sf)
library(ggVennDiagram)
ellipse() \%>\% st_linestring() \%>\% plot()
}
