% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{get_shape_data}
\alias{get_shape_data}
\title{get applicable shape data for Venn object}
\usage{
get_shape_data(nsets, ...)
}
\arguments{
\item{nsets}{number of sets}

\item{...}{
  Arguments passed on to \code{\link[=process_data]{process_data}}
  \describe{
    \item{\code{venn}}{a Venn object}
  }}
}
\value{
a tibble describing specific shape
}
\description{
get applicable shape data for Venn object
}
\examples{
get_shape_data(nsets = 3, type == "polygon")

}
