% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_shape.R
\name{get_shape_data}
\alias{get_shape_data}
\title{get applicable shape data for Venn object}
\usage{
get_shape_data(nsets, type = NULL, shape_id = NULL)
}
\arguments{
\item{nsets}{number of sets}

\item{type}{type of shape}

\item{shape_id}{shape id}
}
\value{
a tibble describing specific shape
}
\description{
ggVennDiagram stores shapes as internal data. You may see all the shapes by
using `plot_shapes()` or `get_shapes()`.
}
\examples{
get_shape_data(nsets = 4, type = "polygon")
}
