% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{ggVennDiagram}
\alias{ggVennDiagram}
\title{ggVennDiagram}
\usage{
ggVennDiagram(x, n.sides = 3000, label = NULL, ...)
}
\arguments{
\item{x}{list of items}

\item{label}{select one from c("count","percent","both")}

\item{...}{}

\item{x}{}
}
\value{
A ggplot object
}
\description{
ggVennDiagram
}
