% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape_generator.R
\name{fancy_4d_ellipse_label}
\alias{fancy_4d_ellipse_label}
\alias{fancy_3d_circle_label}
\alias{fancy_2d_circle_label}
\alias{fancy_6d_triangle_label}
\alias{label_position}
\title{helper function to set label position}
\usage{
fancy_4d_ellipse_label(position = NULL)

fancy_3d_circle_label(position = NULL)

fancy_2d_circle_label(position = NULL)

fancy_6d_triangle_label(position = NULL)

label_position(position)
}
\arguments{
\item{position}{a data.frame containing label coordinates}
}
\value{
a list of matrix
}
\description{
helper function to set label position
}
\details{
\itemize{
\item \code{label_position}: basal wrapper for label postion
\item \code{fancy_6d_triangle_label}: 6 sets triangle label position work with \code{fancy_6d_triangle}
\item \code{fancy_4d_ellipse_label}: 4 sets ellipse label position work with \code{fancy_4d_ellipse}
\item \code{fancy_3d_circle_label}: 3 sets circle label position work with \code{fancy_3d_circle}
\item \code{fancy_2d_circle_label}: 2 sets circle label position work with \code{fancy_2d_circle}
}
}
\examples{
fancy_4d_ellipse_label()
fancy_2d_circle_label()
}
