% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{plot_venn}
\alias{plot_venn}
\title{plot codes}
\usage{
plot_venn(
  x,
  show_intersect,
  label,
  label_geom,
  label_alpha,
  lty,
  percent_digit = 0,
  txtWidth = 40,
  ...
)
}
\arguments{
\item{x}{list of items}

\item{show_intersect}{whether add a hidden text to polygons in the plot, the text can be further visualized by `plotly::ggplotly()`}

\item{label}{select one from c("count","percent","both"), set to NULL if you don't want label.}

\item{label_geom}{choose from geom_label and geom_text}

\item{label_alpha}{set 0 to remove label background}

\item{lty}{line type of polygons}

\item{percent_digit}{number of digits when formating percent label (0)}

\item{txtWidth}{width of text used in showing intersect members (40)}

\item{...}{Other arguments passed on to the polygon layer.}
}
\value{
ggplot object, or plotly object if show_intersect is TRUE
}
\description{
plot codes
}
