% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{plot_venn}
\alias{plot_venn}
\title{reuse plot codes}
\usage{
plot_venn(region_data, category, counts, label, ...)
}
\arguments{
\item{region_data}{a list of two dataframes, which were used to plot polygon and label latter.}

\item{category}{name of Set}

\item{counts}{counts of items for every combinations}

\item{label}{select one from c("count","percent","both")}

\item{...}{Other arguments passed on to the polygon layer.}
}
\value{
ggplot object
}
\description{
reuse plot codes
}
