% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggVennDiagram.R
\name{plot_venn}
\alias{plot_venn}
\title{plot codes}
\usage{
plot_venn(
  data,
  show_intersect = FALSE,
  set_color = "black",
  set_size = NA,
  label = "both",
  label_geom = "label",
  label_alpha = 0.5,
  label_color = "black",
  label_size = NA,
  label_percent_digit = 0,
  label_txtWidth = 40,
  edge_lty = "solid",
  edge_size = 1,
  ...
)
}
\arguments{
\item{data}{plot data}

\item{show_intersect}{if TRUE the text can be visualized by `plotly`}

\item{set_color}{color of set labels ("black")}

\item{set_size}{size of set labels (NA)}

\item{label}{format of region labels, select one from c("count","percent","both","none")}

\item{label_geom}{layer of region labels, choose from c("label", "text")}

\item{label_alpha}{set 0 to remove the background of region labels}

\item{label_color}{color of region labels ("black")}

\item{label_size}{size of region labels (NA)}

\item{label_percent_digit}{number of digits when formatting percent label (0)}

\item{label_txtWidth}{width of text used in showing intersect members, will be ignored unless show_intersection is TRUE (40)}

\item{edge_lty}{line type of set edges ("solid")}

\item{edge_size}{line width of set edges (1)}

\item{...}{useless}
}
\value{
ggplot object, or plotly object if show_intersect is TRUE
}
\description{
plot codes
}
