% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Construct.R
\name{Polygon,ANY-method}
\alias{Polygon,ANY-method}
\alias{polygon}
\title{Polygon constructor}
\usage{
\S4method{Polygon}{ANY}(sets)
}
\arguments{
\item{sets}{a list containing multiple simple features}
}
\description{
Polygon constructor
}
