% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\docType{methods}
\name{process_data}
\alias{process_data}
\alias{process_data,Venn-method}
\title{get plot data}
\usage{
process_data(venn, ...)

\S4method{process_data}{Venn}(venn, ...)
}
\arguments{
\item{venn}{a Venn object}

\item{...}{apply filter to internal shapes. i.e. shape_id == "601", type == "polygon"}
}
\description{
get plot data
}
\examples{
\dontrun{
 venn <- Venn(list(A=1:3,B=2:5,C=4:8))
 data <- process_data(venn)
}
}
