% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\docType{methods}
\name{process_data}
\alias{process_data}
\alias{process_data,Venn-method}
\title{get plot data}
\usage{
process_data(venn, nsets = NULL, shape_id = NULL, type = NULL)

\S4method{process_data}{Venn}(venn, nsets = length(venn@sets), shape_id = NULL, type = NULL)
}
\arguments{
\item{venn}{a Venn object}

\item{nsets}{This parameter will be set automatically.}

\item{shape_id}{apply filter to internal shapes. i.e. shape_id = "601"}

\item{type}{apply filter to internal shapes. i.e. type = "polygon"}
}
\description{
get plot data
}
\details{
This function will conduct set operations and combine the outputs will
stored shapes, thus produce a dataset for plot in next step.

Run `get_shapes()` to show all the characteristics of available shapes.
Run `plot_shapes()` to view those shapes.
}
\examples{
\dontrun{
 venn = Venn(list(A=1:3,B=2:5,C=4:8))
 data = process_data(venn)
}
}
