% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_2d_venn.R, R/draw_3d_venn.R,
%   R/draw_4d_venn.R
\name{two_dimension_circle_regions}
\alias{two_dimension_circle_regions}
\alias{three_dimension_circle_regions}
\alias{region_polygon}
\alias{four_dimension_ellipse_regions}
\title{coordinations of polygon regions/centers for venn diagram}
\usage{
two_dimension_circle_regions(n.sides = 1000)

three_dimension_circle_regions(n.sides = 1000)

four_dimension_ellipse_regions(n.sides = 1000)
}
\arguments{
\item{n.sides}{resolution}
}
\value{
a list of two data.frame, the first one represents polygon regions, and the second specifies polygon centers.
}
\description{
coordinations of polygon regions/centers for venn diagram
}
\examples{
polygons <- four_dimension_ellipse_regions(3000)[[1]]
ggplot(polygons,aes(x,y,group=group,fill=group)) + geom_polygon()
}
