% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_data.R
\name{venn_data}
\alias{venn_data}
\alias{process_set_data}
\alias{process_region_data}
\title{Prepare Venn data}
\usage{
process_set_data(venn)

process_region_data(venn, sep = "/")
}
\arguments{
\item{venn}{a Venn object}

\item{sep}{name and id separator for intersections}
}
\value{
a tibble
}
\description{
Prepare Venn data
}
\examples{
x = list(
A = sample(letters, 8),
B = sample(letters, 8),
C = sample(letters, 8),
D = sample(letters, 8)
)

venn = Venn(x)
process_set_data(venn)
process_region_data(venn)
}
