% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_shape.R
\name{VennPlotData}
\alias{VennPlotData}
\title{An S3 class constructor of representing Venn plot components.}
\usage{
VennPlotData(x)
}
\arguments{
\item{x}{data source of a VennPlotData object}
}
\description{
An S3 class constructor of representing Venn plot components.
}
\section{Slots}{

\describe{
\item{\code{shapeId}}{shape id}

\item{\code{type}}{type of shape}

\item{\code{nsets}}{number of sets}

\item{\code{setEdge}}{a list of coordinates matrix defining Venn set edges}

\item{\code{setLabel}}{label of sets}

\item{\code{regionEdge}}{the feature region will be calculated automatically with `setEdge`}

\item{\code{regionLabel}}{the centroid of region, where region label appears}
}}

