% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_shape.R
\name{VennPlotData}
\alias{VennPlotData}
\title{An S3 class constructor of representing Venn plot components.}
\usage{
VennPlotData(x)
}
\arguments{
\item{x}{data source of a VennPlotData object}
}
\description{
An S3 class constructor of representing Venn plot components.
}
\section{Slots}{

\describe{
\item{\code{shapeId}}{shape id}

\item{\code{type}}{type of shape}

\item{\code{nsets}}{number of sets}

\item{\code{setEdge}}{a data.frame, the coordinates of set edges, can be retrieved by \code{venn_setedge()}}

\item{\code{setLabel}}{a data.frame, the coordinates of set labels, can be retrieved by \code{venn_setlabel()}}

\item{\code{regionEdge}}{a data.frame, the coordinates of different regions, can be retrieved by \code{venn_regionedge()}}

\item{\code{regionLabel}}{a data.frame, the centroid of the regions, where region labels anchored, can be retrieved by \code{venn_regionlabel()}}

\item{\code{setData}}{a data.frame, the set data provided by user, can be retrieved by \code{venn_set()}}

\item{\code{regionData}}{a data.frame, the region data that calculated by \code{ggVennDiagram}, can be retrieved by \code{venn_region()}}
}}

